Changed search and added geospatial + buttons
1. Refined the search for keywords. Instead of using s.search(), and
having to specify size = 9999999..., used helpers.scan() to do the batch
job. It is more efficient for large queries

2. Added geospatial search. Click on map and it will show only the
tweets within 100mi of the point clicked.

Note: all the old code is at the bottom of each file, commented out, so
that we can compare the difference if anything goes wrong.

Important thing to note:
AWS ElasticSearch is an older version than the one I used locally. Need
to check that changing from search() to helpers.scan() to fetch tweets
doesn't break it. If it does and we can't get it to work, can always
revert back to the old way.